requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.9
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.27.0
pydantic==2.7.4
python-dotenv==1.0.0
anyio>=3.7.1,<4.0.0
starlette>=0.27.0,<0.42.0
dashscope==1.20.0 








